feat: fix authentication, don't document authentication for now #553
Also requests like "fields=[<primaryKey>]" should be considered to return detailed data.
We decided not to document authentication for now. The current implementation is special logic required for the GISAID instance. It might be removed in the future. Anyway, if someone else need authentication on LAPIS, we should implement something better.