feat(lapis): add advanced queries feature (#1144)
BREAKING CHANGE: lapisAllowsRegexSearch on string metadata fields in the database config is not necessary anymore, regex search is always enabled for all string fields.

## Summary
When enabled on a LAPIS instance users can make advanced queries using advancedQuery, e.g.

curl -X 'POST' \
  'https://lapis.<instance>.org/open/v2/sample/details' \
  -H 'accept: application/json' \
  -H 'Content-Type: application/json' \
  -d '{
  "advancedQuery": "'\''[3-of: 123A, 234T, S:345-, ORF1a:456K]"
}'

advancedQuery can be thought of as a superset of variantQueries. It should enable all queries also enabled for variantQueries but work on all organisms, additionally it enables metadata filter queries to be joined with variantQueries. In order to do this it must read in the config - the information from the config is only available at runtime and can thus not be passed to the ANTLR grammar parser. Therefore, config-based checks e.g. valid gene name, valid metadata field name and type are done in the backend after the ANTLR parser has been applied to the query.

### Special cases:

currently this does not custom support for specific lineage systems and just assumes that lineages ending in * should include all sublineages.

MAYBE() is only enabled for variant-related queries (I checked with @Taepper this should only work on sequence filter fields. SILO accepts MAYBE requests with metadata filters but handles metadata filters inside a MAYBE as exact matches)

As discussed dateTo and dateFrom queries (also for int and floats) are no longer allowed and date >= and date <= must be used instead.

---------

Co-authored-by: Fabian Engelniederhammer <92720311+fengelniederhammer@users.noreply.github.com>